feat: apm应用列表抽取 # Reviewed, transaction id: 18571
