+++ bkmonitor/webpack/src/apm/pages/home/apm-home.scss
@@ -229,8 +229,8 @@
             padding-bottom: 13px;
             background-color: #fff;
 
-            .title {
-              font-size: 16px;
+            .folding {
+              font-size: 22px;
             }
           }
 
@@ -245,13 +245,13 @@
 
         .toggle-wrap {
           margin-left: 6px;
-          font-size: 24px;
+          font-size: 22px;
           color: #979ba5;
           cursor: pointer;
           transform: rotate(90deg);
 
           .icon-double-up {
-            font-size: 24px;
+            font-size: 22px;
           }
         }
 
@@ -309,11 +309,11 @@
         flex-direction: column;
         padding-top: 16px;
 
-        .item-expan-wrap {
+        .item-expand-wrap {
           margin-bottom: 16px;
           cursor: pointer;
 
-          .expan-header {
+          .expand-header {
             display: flex;
             align-items: center;
             height: 52px;
@@ -332,7 +332,7 @@
                 transition: 0.3s transform;
                 transform: rotate(-90deg);
 
-                &.expan {
+                &.expand {
                   color: #63656e;
                   transform: rotate(0);
                 }
@@ -340,7 +340,7 @@
             }
           }
 
-          .expan-content {
+          .expand-content {
             min-height: 180px;
             background: #fff;
